arg parse generator

+++ nlse_generator.py
@@ -161,7 +161,7 @@ def data_augmentation(
     lines = [20, 50, 100]
     augmentation = len(noises) + len(lines) * len(noises) * len(angles) + 1
 
-    augmented_data = np.zeros((augmentation*E.shape[0], E.shape[1], E.shape[2],E.shape[3]), dtype=np.float16)
+    augmented_data = np.zeros((augmentation*E.shape[0], E.shape[1], E.shape[2],E.shape[3]))
   
     for channel in range(E.shape[1]):
         index = 0
@@ -178,9 +178,9 @@ def data_augmentation(
                         index += 1
     if not (path == None):
         if power != 0:
-            np.save(f'{path}/Es_w{augmented_data.shape[-1]}_n2{number_of_n2}_Isat{number_of_isat}_power{1}_at{str(power)[:4]}_amp_pha_pha_unwrap_extended', augmented_data)
+            np.save(f'{path}/Es_w{augmented_data.shape[-1]}_n2{number_of_n2}_Isat{number_of_isat}_power{1}_at{str(power)[:4]}_amp_pha_pha_unwrap_extended', augmented_data.astype(np.float16))
         else:
-            np.save(f'{path}/Es_w{augmented_data.shape[-1]}_n2{number_of_n2}_Isat{number_of_isat}_power{number_of_power}_amp_pha_pha_unwrap_all_extended', augmented_data)
+            np.save(f'{path}/Es_w{augmented_data.shape[-1]}_n2{number_of_n2}_Isat{number_of_isat}_power{number_of_power}_amp_pha_pha_unwrap_all_extended', augmented_data.astype(np.float16))
     return augmented_data, augmentation
 
 def normalize_data(
